[DEV-6440] duplicate code removal
