+++ usaspending_api/reporting/v2/views/agencies/agency_code/overview.py
@@ -1,21 +1,28 @@
 from django.db.models import Subquery, OuterRef, DecimalField, Func, F, Q, IntegerField
 from rest_framework.response import Response
-from usaspending_api.agency.v2.views.agency_base import AgencyBase
-from django.utils.functional import cached_property
+from usaspending_api.agency.v2.views.agency_base import AgencyBase, PaginationMixin
 
-from usaspending_api.common.data_classes import Pagination
 from usaspending_api.common.helpers.generic_helper import get_pagination_metadata
-from usaspending_api.common.validator import customize_pagination_with_sort_columns, TinyShield
 from usaspending_api.reporting.models import ReportingAgencyOverview, ReportingAgencyTas, ReportingAgencyMissingTas
 from usaspending_api.submissions.models import SubmissionAttributes
 
 
-class AgencyOverview(AgencyBase):
+class AgencyOverview(AgencyBase, PaginationMixin):
     """Returns an overview of the specified agency's submission data"""
 
     endpoint_doc = "usaspending_api/api_contracts/contracts/v2/reporting/agencies/agency_code/overview.md"
 
     def get(self, request, toptier_code):
+        self.sortable_columns = [
+            "fiscal_year",
+            "current_total_budget_authority_amount",
+            "missing_tas_accounts_total",
+            "obligation_difference",
+            "recent_publication_date",
+            "recent_publication_date_certified",
+            "tas_obligation_not_in_gtas_total",
+        ]
+        self.default_sort_column = "current_total_budget_authority_amount"
         results = self.get_agency_overview()
         page_metadata = get_pagination_metadata(len(results), self.pagination.limit, self.pagination.page)
         results = results[self.pagination.lower_limit : self.pagination.upper_limit]
@@ -113,26 +120,3 @@ class AgencyOverview(AgencyBase):
             reverse=self.pagination.sort_order == "desc",
         )
         return results
-
-    @cached_property
-    def pagination(self):
-        sortable_columns = [
-            "fiscal_year",
-            "current_total_budget_authority_amount",
-            "missing_tas_accounts_total",
-            "obligation_difference",
-            "recent_publication_date",
-            "recent_publication_date_certified",
-            "tas_obligation_not_in_gtas_total",
-        ]
-        default_sort_column = "current_total_budget_authority_amount"
-        model = customize_pagination_with_sort_columns(sortable_columns, default_sort_column)
-        request_data = TinyShield(model).block(self.request.query_params)
-        return Pagination(
-            page=request_data["page"],
-            limit=request_data["limit"],
-            lower_limit=(request_data["page"] - 1) * request_data["limit"],
-            upper_limit=(request_data["page"] * request_data["limit"]),
-            sort_key=request_data.get("sort", default_sort_column),
-            sort_order=request_data["order"],
-        )
